requirements.txt
# 核心框架
langchain
langchain-core
langchain-openai
langgraph
langchain-community
langchain-experimental # 修复：为PythonREPLTool添加

# 模型与嵌入
sentence-transformers
faiss-cpu

# 网页处理与抓取
beautifulsoup4
lxml
readability-lxml
requests

# 配置文件与环境变量
pyyaml
python-dotenv
jinja2

# 交互式CLI
InquirerPy

# 搜索引擎API
duckduckgo-search
arxiv
tavily-python

# 文件输出
aiofiles
markdown2
python-pptx
pdfkit
bs4
gtts # 新增：用于生成播客音频

# 异步HTTP客户端
httpx
openai
python-dotenv